[Renamed] type_store files -> reflect_factory [Add] helpers for making Instance refs
