+++ engine/reflection/private/member/field_impl.hpp
@@ -1,8 +1,8 @@
 #pragma once
 #include <member/field.hpp>
 
+#include <factory/reflect_factory.hpp>
 #include <instance/instance.hpp>
-#include <type/store.hpp>
 
 template <typename Class, typename Member>
 Field::Field(TypeStore& type_store, Member Class::* mem_ptr)
